cleanup a bit code + added circles + polylines on map

+++ pages/configuration/mission/php/deleteMissionFromDB.php
@@ -1,6 +1,7 @@
 <?php
 define('__ROOT__', dirname(dirname(dirname(dirname(dirname(__FILE__))))));
 require_once(__ROOT__.'/globalsettings.php');
+require_once('is_ajax.php');
 
 function deleteMissionFromDB($id_mission)
 {
@@ -26,17 +27,37 @@ function deleteMissionFromDB($id_mission)
         die('Error : '.$e->getMessage());
     }
 
-    $query = $db->prepare('DELETE FROM mission WHERE mission.id=?;');
-    $query->execute(array($id_mission));
-    $query = $db->prepare('DELETE FROM pointList WHERE pointList.id_mission=?;');
-    $query->execute(array($id_mission));
+    // Delete the mission
+    $query1 = $db->prepare('DELETE FROM mission WHERE mission.id=?;');
+    $exec1 = $query1->execute(array(htmlspecialchars($id_mission)));
+    
+    // Delete its waypoints and checkpoints
+    $query2 = $db->prepare('DELETE FROM pointList WHERE pointList.id_mission=?;');
+    $exec2 = $query2->execute(array(htmlspecialchars($id_mission)));
 
-    $result = 1;
-    return $result;
+    // Check if success or not
+    if ($exec1 == false)
+    {
+        $msg = sprintf("Error while deleting mission from DB (website) because execute() failed: %s\n<br />", htmlspecialchars($query1->error));
+    }
+    if ($exec2 == false)
+    {
+        $msg = sprintf("Error while deleting old pointList from DB (website) because execute() failed: %s\n<br />", htmlspecialchars($query2->error));
+    }
+    if ($exec1 && $exec2)
+    {
+        $msg = "Success !";
+    }
+
+    echo $msg;
+    
+    // End the queries properly
+    $query1->closeCursor();
+    $query2->closeCursor();
 }
 
 session_start();
-if (isset($_POST['id_mission']) && $_SESSION['right'] == 'admin')
+if (is_ajax() && isset($_POST['id_mission']) && $_SESSION['right'] == 'admin')
 {
     deleteMissionFromDB($_POST['id_mission']);
 }
